Added a general buffer region of GBA memory
This commit intends to extend the memory available to GBA ROMs by
utilizing the "unused memory" address range of GBA ROMs. This extension
gives a 15MB~ buffer for GBA ROMs and ROM developers to use as just
extra memory. In our case, this will be utilized to create a
communication layer between ROMs and the emulator in order to support
online capabilities.